i2c: cleanup and refactoring
