+++ drivers/i2c/meson/include/clk.h
@@ -6,7 +6,7 @@
 
 #ifndef __I2C_CLK_H__
 #define __I2C_CLK_H__
-#define I2C_CLK_OFFSET 320   
-#define I2C_CLK81_BIT 0x1 << 9 // bit 9
+#define I2C_CLK_OFFSET 320
+#define I2C_CLK81_BIT (1 << 9) // bit 9
 
 #endif
