+++ examples/i2c/pn532.c
@@ -1,11 +1,11 @@
 #include <stdint.h>
 #include <libco.h>
+#include <sddf/util/printf.h>
 #include "pn532.h"
 #include "client.h"
 
 #include "i2c.h"
-#include "printf.h"
-#include "i2c-transport.h"
+#include "transport.h"
 
 #define DEBUG_PN532
